Fixing bug in create_ds_lib.py

+++ src/create_ds_lib.py
@@ -5,10 +5,13 @@ import shutil
 
 SUBDIRS = ["bst", "dll", "queue", "scll", "sll", "stack"]
 
-shutil.rmtree("bin")
+if os.path.exists("bin"):
+	shutil.rmtree("bin")
+
 os.mkdir("bin")
 
 for subdir in SUBDIRS:
 	os.system("cp -f %s/*.o bin/." % (subdir))
 
 os.system("cp -f main.o bin/.")
+
